Moved get_project_xml_data() from paths.py to parsers.py
