+++ filesystem/tests/paths_test.py
@@ -244,35 +244,22 @@ def test_get_project_root_then_calls_get_file_path_in_tree_with_project_file():
 
 # endregion
 
-# region get_project_xml_data()
+# region is_empty_dir()
 
 
-def test_get_project_xml_data_when_add_environment_variables_is_false_then_return_dict_with_env_variables(paths):
-    """When add_environment_variables is false, then return dict with xml data"""
-
-    # Arrange
-    expected_result = {"PROJECT_FOO1": "foo1", "PROJECT_FOO2": "foo2", "PROJECT_FOO3": "foo3"}
-    with patch.object(pathlib.Path, "joinpath") as joinpath_mock:
-        joinpath_mock.return_value = paths.file_foo_xml_project_path
-        # Act
-        result = sut.get_project_xml_data(False)
-        # Assert
-        assert expected_result == result
-
-
-def test_get_project_xml_data_when_add_environment_variables_is_true_then_call_create_env_variables(paths):
-    """When add_environment_variables is false, then return dict with xml data"""
+def test_is_empty_dir(paths):
+    """Given a directory path, when it is empty, returns True"""
 
     # Arrange
-    expected_result = {"PROJECT_FOO1": "foo1", "PROJECT_FOO2": "foo2", "PROJECT_FOO3": "foo3"}
-    with patch.object(pathlib.Path, "joinpath") as joinpath_mock:
-        joinpath_mock.return_value = paths.file_foo_xml_project_path
-        with patch.object(sut, "platform_specific") as platform_specific_mock:
-            with patch.object(platform_specific_mock, "create_environment_variables") as create_env_vars_mock:
-                # Act
-                sut.get_project_xml_data(True)
-                # Assert
-                create_env_vars_mock.assert_called_once_with(expected_result)
+
+    # Act
+
+    # Assert
+    assert True
+
+# endregion
+
+# region is_valid_path()
 
 
 def test_is_valid_path_given_test_path_returns_true(paths):
@@ -299,20 +286,5 @@ def test_is_valid_path_given_non_existent_path_returns_false(paths):
 
     # Assert
     assert not result
-
-# endregion
-
-# region is_empty_dir()
-
-
-def test_is_empty_dir(paths):
-    """Given a directory path, when it is empty, returns True"""
-
-    # Arrange
-
-    # Act
-
-    # Assert
-    assert True
 
 # endregion
